Project-Complaint_Management_System/Source Code/requirements.txt
Flask==2.3.3
pymongo==4.5.0
bcrypt==4.0.1
python-dotenv==1.0.0